+++ src/database/migrations/2021_01_11_164251_add_short_name_to_partner.php
@@ -1,0 +1,31 @@
+<?php
+
+use Illuminate\Database\Migrations\Migration;
+use Illuminate\Support\Facades\Schema;
+
+class AddShortNameToPartner extends Migration
+{
+    /**
+     * Run the migrations.
+     *
+     * @return void
+     */
+    public function up()
+    {
+        Schema::table('partners', function ($table) {
+            $table->string('short_name')->nullable();
+        });
+    }
+
+    /**
+     * Reverse the migrations.
+     *
+     * @return void
+     */
+    public function down()
+    {
+        Schema::table('partners', function ($table) {
+            $table->dropColumn('short_name');
+        });
+    }
+}