+++ src/resources/lang/dwfw/en/partners.php
@@ -4,6 +4,7 @@ return [
     'partners' => 'Partners',
     'partner' => 'Partner',
     'name' => 'Name',
+    'short_name' => 'Short name',
     'contact_name' => 'Contact name',
     'contact_phone' => 'Contact phone',
     'contact_email' => 'Contact email',