+++ src/resources/lang/dwfw/hu/partners.php
@@ -4,6 +4,7 @@ return [
     'partners' => 'Partnerek',
     'partner' => 'Partner',
     'name' => 'Név',
+    'short_name' => 'Rövid név',
     'contact_name' => 'Kapcsolattartó neve',
     'contact_phone' => 'Kapcsolattartó telefonszáma',
     'contact_email' => 'Kapcsolattartó e-mail címe',