+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.11.1] - 2021-01-11
+###changed
+-partner with short name
+-required php version changed to 7.4 or greater
+
 ## [0.11.0] - 2021-01-06
 ###changed
 -checkForColumnId with return value